+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2026-04-16
+
+### Fixed
+
+- Fixed chain staff recipe on 1.20.
+- Added chain and zipline staffs to the creative menu when enabled.
+
 ## [1.3.1] - 2026-04-13
 
 ### Fixed